+++ src/app/error/api.ts
@@ -6,9 +6,9 @@ import { factory } from '@motionpicture/sskts-domain';
  */
 export class APIError extends Error {
     public readonly code: number;
-    public readonly errors: factory.errors.SSKTS[];
+    public readonly errors: factory.errors.Cinerino[];
 
-    constructor(code: number, errors: factory.errors.SSKTS[]) {
+    constructor(code: number, errors: factory.errors.Cinerino[]) {
         const message = errors.map((error) => error.message).join('\n');
         super(message);